refactored space-weather-notifications and modal
